requirements.txt
pyTelegramBotAPI==4.22.1
jupyter~=1.0.0
style~=1.1.0
pip~=23.3.1
attrs~=23.1.0
wheel~=0.38.4
cryptography~=41.0.5
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.2
certifi~=2023.7.22
fastapi~=0.92.0
uvicorn~=0.20.0
psutil~=5.9.4
watchdog~=3.0.0
redis~=5.0.3
ujson~=5.8.0
requests~=2.31.0
aiohttp~=3.8.3
zstandard~=0.17.0
urllib3~=2.0.7
idna~=3.4
setuptools~=65.5.0
packaging~=22.0
pyparsing~=3.1.1
python-dotenvaiohttp==3.8.6
aiosignal==1.3.1
anyio==4.5.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==41.0.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastapi==0.92.0
fastjsonschema==2.20.0
fqdn==1.5.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
overrides==7.7.0
packaging==22.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==1.10.18
Pygments==2.18.0
pyparsing==3.1.4
pyTelegramBotAPI==4.22.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.0
QtPy==2.4.1
redis==5.0.8
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.25.0
style==1.1.6
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yarl==1.11.1
zstandard==0.17.0
